+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 3.4.1
 - Fix Remove previous years MP columns from list in Transcripts.php
 - Fix SQL error remove duplicate "s." prefix in Search.fnc.php
 - Remove useless DBGet indexes in SetUserStudentID() in Current.php
+- Fix PHP error when no Student associated to Parent in Registration.php
 
 Changes in 3.4
 --------------